projekt-1-pzx106184/requirements.txt
pyopengl
pyglm
glfw
numpy